feat: move server modules out of client components
